user docs: Remove right simplebar from help and api page.
This commit removes the simplebar implementation for the right on /help.
Thus using the default browser scrollbar. Removes the simplebar class and
the function scrollToHash with this commit.

Currently the markdown content handles the scroll automatically using the
broweser scrollbar. Also manged to open hash links automatically and
scrolls if necessay. As the update page is called, it also checks if the
hash is empty or not. If the hash is empty, scroll it to top forcefully
otherwise the scrollbar automatically handles the hash location and scroll.

Fixes #10732