user docs: Converted browser scrollbar overflow from .markdown to html.
This commit removes the overflow hidden from the body and overflow auto
from the .markdown element. After this, it adds a scroll margin top to
the heading elements so that using the hash, the page opens at perfect
location. It adds position fixed to the .sidebar so that it remains fixes
and the html scroll does not make it scroll. This all ensures that the
scrollbar works perfectly and scrolls the page properly. It also removes
the default height set by the body to 100% and the default height set to
.markdown element. Without that, the height of the elements get weird
and this can result in coming of scrollbar wherever not neccessary.

For the mobile view, it changes the sidebar right from 100vw to 100%.
This also ensures that the page works perfectly for the mobile devices.

This also changes the way how the hamburger menu changes its position.
It uses right propery now instead of left as it works better than
translate combined with left.

m