frontend: Define a dark color palette for consistency.
This commit replaces the colors with pallets.
Will update the commit message depending on the feedback.

It changes the hover/selected elements color to color-dark-night.
It changes some elements like table header, outlines with darker.
And more darker ones with darkest.

Some finising in dark theme like different color values for almost
same color. So, replacing them with proper color that they need from
the palette that is created.